Rename folder to intermediate name for case sensitivity fix
